Reorganize so that similar functions are together, as in MindtPy
